implement regal based rego file validation
